chore: refactor tailnetAPIConnector to tailnet.Controller
