+++ tailnet/controllers_test.go
@@ -10,6 +10,8 @@ import (
 	"time"
 
 	"github.com/google/uuid"
+	"github.com/hashicorp/yamux"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/mock/gomock"
 	"golang.org/x/xerrors"
@@ -677,4 +679,203 @@ func (f *fakeResumeTokenClient) RefreshResumeToken(_ context.Context, _ *proto.R
 type fakeResumeTokenCall struct {
 	resp  chan *proto.RefreshResumeTokenResponse
 	errCh chan error
+}
+
+func TestController_Disconnects(t *testing.T) {
+	t.Parallel()
+	testCtx := testutil.Context(t, testutil.WaitShort)
+	ctx, cancel := context.WithCancel(testCtx)
+	logger := slogtest.Make(t, &slogtest.Options{
+		IgnoredErrorIs: append(slogtest.DefaultIgnoredErrorIs,
+			io.EOF,                   // we get EOF when we simulate a DERPMap error
+			yamux.ErrSessionShutdown, // coordination can throw these when DERP error tears down session
+		),
+	}).Leveled(slog.LevelDebug)
+	agentID := uuid.UUID{0x55}
+	clientID := uuid.UUID{0x66}
+	fCoord := tailnettest.NewFakeCoordinator()
+	var coord tailnet.Coordinator = fCoord
+	coordPtr := atomic.Pointer[tailnet.Coordinator]{}
+	coordPtr.Store(&coord)
+	derpMapCh := make(chan *tailcfg.DERPMap)
+	defer close(derpMapCh)
+	svc, err := tailnet.NewClientService(tailnet.ClientServiceOptions{
+		Logger:                  logger.Named("svc"),
+		CoordPtr:                &coordPtr,
+		DERPMapUpdateFrequency:  time.Millisecond,
+		DERPMapFn:               func() *tailcfg.DERPMap { return <-derpMapCh },
+		NetworkTelemetryHandler: func([]*proto.TelemetryEvent) {},
+		ResumeTokenProvider:     tailnet.NewInsecureTestResumeTokenProvider(),
+	})
+	require.NoError(t, err)
+
+	dialer := &pipeDialer{
+		ctx:    testCtx,
+		logger: logger,
+		t:      t,
+		svc:    svc,
+		streamID: tailnet.StreamID{
+			Name: "client",
+			ID:   clientID,
+			Auth: tailnet.ClientCoordinateeAuth{AgentID: agentID},
+		},
+	}
+
+	peersLost := make(chan struct{})
+	fConn := &fakeTailnetConn{peersLostCh: peersLost}
+
+	uut := tailnet.NewController(logger.Named("tac"), dialer,
+		// darwin can be slow sometimes.
+		tailnet.WithGracefulTimeout(5*time.Second))
+	uut.CoordCtrl = tailnet.NewAgentCoordinationController(logger.Named("coord_ctrl"), fConn)
+	uut.DERPCtrl = tailnet.NewBasicDERPController(logger.Named("derp_ctrl"), fConn)
+	uut.Run(ctx)
+
+	call := testutil.RequireRecvCtx(testCtx, t, fCoord.CoordinateCalls)
+
+	// simulate a problem with DERPMaps by sending nil
+	testutil.RequireSendCtx(testCtx, t, derpMapCh, nil)
+
+	// this should cause the coordinate call to hang up WITHOUT disconnecting
+	reqNil := testutil.RequireRecvCtx(testCtx, t, call.Reqs)
+	require.Nil(t, reqNil)
+
+	// and mark all peers lost
+	_ = testutil.RequireRecvCtx(testCtx, t, peersLost)
+
+	// ...and then reconnect
+	call = testutil.RequireRecvCtx(testCtx, t, fCoord.CoordinateCalls)
+
+	// canceling the context should trigger the disconnect message
+	cancel()
+	reqDisc := testutil.RequireRecvCtx(testCtx, t, call.Reqs)
+	require.NotNil(t, reqDisc)
+	require.NotNil(t, reqDisc.Disconnect)
+	close(call.Resps)
+
+	_ = testutil.RequireRecvCtx(testCtx, t, peersLost)
+}
+
+func TestController_TelemetrySuccess(t *testing.T) {
+	t.Parallel()
+	ctx := testutil.Context(t, testutil.WaitShort)
+	logger := slogtest.Make(t, nil).Leveled(slog.LevelDebug)
+	agentID := uuid.UUID{0x55}
+	clientID := uuid.UUID{0x66}
+	fCoord := tailnettest.NewFakeCoordinator()
+	var coord tailnet.Coordinator = fCoord
+	coordPtr := atomic.Pointer[tailnet.Coordinator]{}
+	coordPtr.Store(&coord)
+	derpMapCh := make(chan *tailcfg.DERPMap)
+	defer close(derpMapCh)
+	eventCh := make(chan []*proto.TelemetryEvent, 1)
+	svc, err := tailnet.NewClientService(tailnet.ClientServiceOptions{
+		Logger:                 logger,
+		CoordPtr:               &coordPtr,
+		DERPMapUpdateFrequency: time.Millisecond,
+		DERPMapFn:              func() *tailcfg.DERPMap { return <-derpMapCh },
+		NetworkTelemetryHandler: func(batch []*proto.TelemetryEvent) {
+			select {
+			case <-ctx.Done():
+				t.Error("timeout sending telemetry event")
+			case eventCh <- batch:
+				t.Log("sent telemetry batch")
+			}
+		},
+		ResumeTokenProvider: tailnet.NewInsecureTestResumeTokenProvider(),
+	})
+	require.NoError(t, err)
+
+	dialer := &pipeDialer{
+		ctx:    ctx,
+		logger: logger,
+		t:      t,
+		svc:    svc,
+		streamID: tailnet.StreamID{
+			Name: "client",
+			ID:   clientID,
+			Auth: tailnet.ClientCoordinateeAuth{AgentID: agentID},
+		},
+	}
+
+	uut := tailnet.NewController(logger, dialer)
+	uut.CoordCtrl = tailnet.NewAgentCoordinationController(logger, &fakeTailnetConn{})
+	tel := tailnet.NewBasicTelemetryController(logger)
+	uut.TelemetryCtrl = tel
+	uut.Run(ctx)
+	// Coordinate calls happen _after_ telemetry is connected up, so we use this
+	// to ensure telemetry is connected before sending our event
+	cc := testutil.RequireRecvCtx(ctx, t, fCoord.CoordinateCalls)
+	defer close(cc.Resps)
+
+	tel.SendTelemetryEvent(&proto.TelemetryEvent{
+		Id: []byte("test event"),
+	})
+
+	testEvents := testutil.RequireRecvCtx(ctx, t, eventCh)
+
+	require.Len(t, testEvents, 1)
+	require.Equal(t, []byte("test event"), testEvents[0].Id)
+}
+
+type fakeTailnetConn struct {
+	peersLostCh chan struct{}
+}
+
+func (*fakeTailnetConn) UpdatePeers([]*proto.CoordinateResponse_PeerUpdate) error {
+	// TODO implement me
+	panic("implement me")
+}
+
+func (f *fakeTailnetConn) SetAllPeersLost() {
+	if f.peersLostCh == nil {
+		return
+	}
+	f.peersLostCh <- struct{}{}
+}
+
+func (*fakeTailnetConn) SetNodeCallback(func(*tailnet.Node)) {}
+
+func (*fakeTailnetConn) SetDERPMap(*tailcfg.DERPMap) {}
+
+func (*fakeTailnetConn) SetTunnelDestination(uuid.UUID) {}
+
+type pipeDialer struct {
+	ctx      context.Context
+	logger   slog.Logger
+	t        testing.TB
+	svc      *tailnet.ClientService
+	streamID tailnet.StreamID
+}
+
+func (p *pipeDialer) Dial(_ context.Context, _ tailnet.ResumeTokenController) (tailnet.ControlProtocolClients, error) {
+	s, c := net.Pipe()
+	go func() {
+		err := p.svc.ServeConnV2(p.ctx, s, p.streamID)
+		p.logger.Debug(p.ctx, "piped tailnet service complete", slog.Error(err))
+	}()
+	client, err := tailnet.NewDRPCClient(c, p.logger)
+	if !assert.NoError(p.t, err) {
+		_ = c.Close()
+		return tailnet.ControlProtocolClients{}, err
+	}
+	coord, err := client.Coordinate(context.Background())
+	if !assert.NoError(p.t, err) {
+		_ = c.Close()
+		return tailnet.ControlProtocolClients{}, err
+	}
+
+	derps := &tailnet.DERPFromDRPCWrapper{}
+	derps.Client, err = client.StreamDERPMaps(context.Background(), &proto.StreamDERPMapsRequest{})
+	if !assert.NoError(p.t, err) {
+		_ = c.Close()
+		return tailnet.ControlProtocolClients{}, err
+	}
+	return tailnet.ControlProtocolClients{
+		Closer:      client.DRPCConn(),
+		Coordinator: coord,
+		DERP:        derps,
+		ResumeToken: client,
+		Telemetry:   client,
+	}, nil
 }
